backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==3.2
django-extensions==3.2.3
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
idna==3.10
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
oauthlib==3.2.2
Pillow==9.3.0
psycopg2-binary==2.9.3
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
python-dotenv==1.0.1
